html/parsing/html5lib/html-tree/tests21.dat

#data
<svg><![CDATA[foo]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<math><![CDATA[foo]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       "foo"

#data
<div><![CDATA[foo]]>
#errors
no DOCTYPE
bogus comment
EOF
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[foo
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[foo
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "foo"

#data
<svg><![CDATA[
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]] >]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]] >]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]] >"

#data
<svg><![CDATA[]]
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]]"

#data
<svg><![CDATA[]
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]"

#data
<svg><![CDATA[]>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "]>a"

#data
<svg><foreignObject><div><![CDATA[foo]]>
#errors
no DOCTYPE
bogus comment
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <div>
|           <!-- [CDATA[foo]] -->

#data
<svg><![CDATA[<svg>]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[</svg>a]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>a"

#data
<svg><![CDATA[<svg>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>a"

#data
<svg><![CDATA[</svg>a
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "</svg>a"

#data
<svg><![CDATA[<svg>]]><path>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <svg path>

#data
<svg><![CDATA[<svg>]]></path>
#errors
no DOCTYPE
foreign </path>
unmatched </path>
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"

#data
<svg><![CDATA[<svg>]]><!--path-->
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>"
|       <!-- path -->

#data
<svg><![CDATA[<svg>]]>path
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<svg>path"

#data
<svg><![CDATA[<!--svg-->]]>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "<!--svg-->"
